chore(split-in-components): split app in two components WIP
